Convert controlsvc subcommands to an app
